requirements.txt
boto3==1.9.215
python-dotenv==0.10.3
